Use normalized inputs for the neural network to work better with sigmoid function

+++ app/src/main/java/com/ps/simplepapersoccer/AI/neuralnetworkAI/INeuralNetworkController.kt
@@ -1,7 +1,7 @@
 package com.ps.simplepapersoccer.ai.neuralnetworkAI
 
 interface INeuralNetworkController<T> {
-    val inputs: List<Int>
+    val inputs: List<Double>
     val outputs: Int
     fun fitnessEvaluation(): Double
     fun networkGuessOutput(output: List<Double>): T?
